backend/requirements.txt
django>=5.1,<6.1
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
stripe>=8.0,<12
twilio>=9.0,<10
python-dotenv>=1.0,<2
